saa716x_ff: Adapted frontend init to latest driver changes.
With latest changes on stv090x and stv6110x drivers regardings the tuner I2C gate locking it is needed to initialize the demodulator directly after attaching it. This leads to a correct setup of the tuner's clock output divider. In case of the TT S2-6400 we must initialize adapter 1 as the second tuner drives the STV0900.

Signed-off-by: Andreas Regel <andreas.regel@gmx.de>