saa716x_ff: Don't use dvb_generic_ioctl.
As dvb_generic_ioctl uses dvb-global locking since october 2010, which
is not needed for the saa716x_ff, support for using it has been
removed. Instead local ioctl functions are used that do the usercopy
internally.

This has the positive effect that an access to the firmware e.g. drawing
to the OSD is no longer slowed down by a frontend access, especially
during bad reception conditions.

Signed-off-by: Andreas Regel <andreas.regel@gmx.de>

[moved from media/common/saa716x/ to staging/media/saa716x/]
Signed-off-by: Soeren Moch <smoch@web.de>