saa716x_ff: Support capturing digital video coming from STi7109.
The decoder chipset (STi7109) outputs HD video through its DVO port. This
video contains the same picture that is visible on the HDMI output. With the
latest changes it is possible to capture that video using one of the VI units
of the SAA716x. There are some limitations though:

- The resolution is the same as the signal on the HDMI.
- Only formats 720p50/60, 1080i50/60 and 1080p24/25/30 are supported. 576p and
  576i formats lead to strange behaviour of the SAA716x. So if the HDMI uses
  that formats, it will not be possible to capture the video.
- To keep the required bandwidth low, only one-shot capturing is used.
- The current implementation does not use the V4L interface (/dev/videoX) to
  provide the captured pictures to user space. Instead it is available in
  proprietary form through the DVB video device: A v4l2_pix_format header
  followed by the picture data in UYVY format.

Some of them might be resolved later.

The support is activated through the "video_capture" module parameter.

Signed-off-by: Andreas Regel <andreas.regel@gmx.de>