saa716x_ff: Cancel data transfer in case of a block timeout
In case a block timeout occured during a data transfer, e. g. because
of high system load, no further data transfers could be done until the
driver was reloaded. Now an explicit data interrupt is sent to the
STi7109 which triggers a cancellation of the ongoing transfer, so
further transfers will be possible.

This feature requires firmware 0.3.10 or later. Firmware versions up to
and including 0.3.9 will just ignore the interrupt.

Signed-off-by: Andreas Regel <andreas.regel@gmx.de>

[moved from media/common/saa716x/ to staging/media/saa716x/]
Signed-off-by: Soeren Moch <smoch@web.de>