saa716x_ff: Align block data on 32 bytes for firmware 0.5.0 and up
Firmware 0.5.0 and up support a special DMA transfer for copying block data
that is much faster than the usual one but requires 32 byte aligned data. So
when this version is available the data is moved to the required offset and
the firmware is signalled to use these transfers.

This change together with a compatible firmware improves the performance of
data transfers depending on PHI mode up to 200 percent.

Signed-off-by: Andreas Regel <andreas.regel@gmx.de>