saa716x_ff: Avoid sleeping in fifo_worker
Since tasklet contexts are shared among different drivers, we should not
sleep in the fifo_worker. Instead introduce a new state TSOUT_STAT_WAIT,
which is entered when the tsout buffer is (almost) empty, e.g. because
all available ts packets are already buffered in the decoder. This is the
normal state when watching live TV.
To improve code readability use tsout buffer level constants.

Signed-off-by: Soeren Moch <smoch@web.de>

[moved from media/common/saa716x/ to staging/media/saa716x/]
Signed-off-by: Soeren Moch <smoch@web.de>